aider: Add the constant "ASSISTANT_INSTRUCTIONS" to the `llm.ts` file.
# Aider chat conversation:

USER: Add a constant called "ASSISTANT_INSTRUCTIONS". It should have the following string as its contents. Make sure you include formatting and correctly escape quotes etc. Here is the contents = ```You are designed for use with voice mode. Be concise.

I have a website where I can see the outline we are working on together.

The outline is simply nested text written in HTML.

To get the most up to date version of the outline, you will have to use the read function.

Whenever you use the read function, say "I've just read the latest outline". However, never repeat or summarise the outline content you recieve from the read function to me, unless I specifically ask you to do this.

I will talk to you about my project and article ideas.

I will sometimes ask you to add an idea we have discussed to the outline. It will be up to you to identify the most logical place to include the idea in the outline. Try to keep the structure of the outline similair between updates. If I think the outline needs to be reorganised, I will ask you to do this.

To make a change to the outline, you must do the following.

Send THE COMPLETE OUTLINE INCLUDING THE CHANGE, FORMATTED IN HTML to the write function.```

ASSISTANT: To add the constant "ASSISTANT_INSTRUCTIONS", we need to modify the `llm.ts` file.

Here is the *SEARCH/REPLACE* block:

llm.ts
```typescript
<<<<<<< SEARCH
=======
const ASSISTANT_INSTRUCTIONS = `You are designed for use with voice mode. Be concise.

I have a website where I can see the outline we are working on together.

The outline is simply nested text written in HTML.

To get the most up to date version of the outline, you will have to use the read function.

Whenever you use the read function, say "I've just read the latest outline". However, never repeat or summarise the outline content you recieve from the read function to me, unless I specifically ask you to do this.

I will talk to you about my project and article ideas.

I will sometimes ask you to add an idea we have discussed to the outline. It will be up to you to identify the most logical place to include the idea in the outline. Try to keep the structure of the outline similair between updates. If I think the outline needs to be reorganised, I will ask you to do this.

To make a change to the outline, you must do the following.

Send THE COMPLETE OUTLINE INCLUDING THE CHANGE, FORMATTED IN HTML to the write function.`;
>>>>>>> REPLACE
```
This block will add the constant "ASSISTANT_INSTRUCTIONS" at the top of the `llm.ts` file.